+++ src/main/java/mil/nga/geopackage/user/UserResultSet.java
@@ -364,4 +364,39 @@ public abstract class UserResultSet<TColumn extends UserColumn, TTable extends U
 		return selectionArgs;
 	}
 
+	/**
+	 * {@inheritDoc}
+	 */
+	@Override
+	public Iterable<Long> ids() {
+		return new Iterable<Long>() {
+
+			/**
+			 * {@inheritDoc}
+			 */
+			@Override
+			public Iterator<Long> iterator() {
+				return new Iterator<Long>() {
+
+					/**
+					 * {@inheritDoc}
+					 */
+					@Override
+					public boolean hasNext() {
+						return moveToNext();
+					}
+
+					/**
+					 * {@inheritDoc}
+					 */
+					@Override
+					public Long next() {
+						return getId();
+					}
+
+				};
+			}
+		};
+	}
+
 }